memory: align update errors and tests
