+++ memory/mysqlvec/service_test.go
@@ -913,6 +913,15 @@ func TestService_UpdateMemory_SoftDelete_RotateMemory_ActiveRowConflict(t *testi
 
 	key := memory.Key{AppName: "app", UserID: "u1", MemoryID: "mem-A"}
 	now := time.Now()
+	targetID := imemory.GenerateMemoryID(
+		&memory.Memory{
+			Memory: "content B",
+			Topics: []string{"topic"},
+			Kind:   memory.KindFact,
+		},
+		key.AppName,
+		key.UserID,
+	)
 	// Load returns original entry with ID "mem-A".
 	mock.ExpectQuery("SELECT memory_id").
 		WithArgs(key.MemoryID, key.AppName, key.UserID).
@@ -924,7 +933,7 @@ func TestService_UpdateMemory_SoftDelete_RotateMemory_ActiveRowConflict(t *testi
 	// BEGIN + pre-check finds target is active → error → ROLLBACK.
 	mock.ExpectBegin()
 	mock.ExpectQuery("SELECT deleted_at IS NULL FROM memories").
-		WithArgs(sqlmock.AnyArg(), key.AppName, key.UserID).
+		WithArgs(targetID, key.AppName, key.UserID).
 		WillReturnRows(sqlmock.NewRows([]string{"active"}).AddRow(true))
 	mock.ExpectRollback()
 
@@ -937,7 +946,7 @@ func TestService_UpdateMemory_SoftDelete_RotateMemory_ActiveRowConflict(t *testi
 		memory.WithUpdateResult(result),
 	)
 	require.Error(t, err)
-	assert.Contains(t, err.Error(), "already active")
+	assert.Equal(t, fmt.Sprintf("memory with id %s already exists", targetID), err.Error())
 	assert.Equal(t, "unchanged", result.MemoryID)
 	require.NoError(t, mock.ExpectationsWereMet())
 }
@@ -1044,7 +1053,7 @@ func TestService_UpdateMemory_HardDelete_RotateMemory_ActiveRowConflict(t *testi
 		memory.WithUpdateResult(result),
 	)
 	require.Error(t, err)
-	assert.Contains(t, err.Error(), "already active")
+	assert.Equal(t, fmt.Sprintf("memory with id %s already exists", targetID), err.Error())
 	assert.Equal(t, "unchanged", result.MemoryID)
 	require.NoError(t, mock.ExpectationsWereMet())
 }
@@ -2067,4 +2076,102 @@ func TestOptions_WithCustomTool_ValidName(t *testing.T) {
 	assert.NotNil(t, opts.toolCreators[memory.SearchToolName])
 	_, exists := opts.enabledTools[memory.SearchToolName]
 	assert.True(t, exists)
+}
+
+func TestService_UpdateMemory_HardDeleteRejectsTombstoneSource(t *testing.T) {
+	tests := []struct {
+		name      string
+		memoryStr string
+	}{
+		{
+			name:      "same identity",
+			memoryStr: "source",
+		},
+		{
+			name:      "rotated identity",
+			memoryStr: "target",
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			db, mock := setupMockDB(t)
+			svc := setupMockService(
+				t,
+				db,
+				mock,
+				WithSkipDBInit(true),
+				WithSoftDelete(false),
+			)
+			t.Cleanup(func() {
+				require.NoError(t, mock.ExpectationsWereMet())
+				_ = svc.Close()
+			})
+
+			key := memory.Key{
+				AppName:  "app",
+				UserID:   "user",
+				MemoryID: "source-id",
+			}
+			mock.ExpectQuery("SELECT memory_id, app_name, user_id.*deleted_at IS NULL").
+				WithArgs(key.MemoryID, key.AppName, key.UserID).
+				WillReturnRows(sqlmock.NewRows(memCols))
+
+			result := &memory.UpdateResult{MemoryID: "unchanged"}
+			err := svc.UpdateMemory(
+				context.Background(),
+				key,
+				tt.memoryStr,
+				nil,
+				memory.WithUpdateResult(result),
+			)
+			require.ErrorContains(t, err, "not found")
+			require.Equal(t, "unchanged", result.MemoryID)
+		})
+	}
+}
+
+func TestService_UpdateMemory_HardDeleteRollsBackWhenSourceBecomesTombstone(
+	t *testing.T,
+) {
+	db, mock := setupMockDB(t)
+	svc := setupMockService(
+		t,
+		db,
+		mock,
+		WithSkipDBInit(true),
+		WithSoftDelete(false),
+	)
+	t.Cleanup(func() {
+		require.NoError(t, mock.ExpectationsWereMet())
+		_ = svc.Close()
+	})
+
+	key := memory.Key{
+		AppName:  "app",
+		UserID:   "user",
+		MemoryID: "source-id",
+	}
+	expectUpdateLoad(mock, key, false)
+	mock.ExpectBegin()
+	mock.ExpectQuery("SELECT deleted_at IS NULL FROM memories.*FOR UPDATE").
+		WithArgs(sqlmock.AnyArg(), key.AppName, key.UserID).
+		WillReturnError(sql.ErrNoRows)
+	mock.ExpectExec("INSERT INTO").
+		WillReturnResult(sqlmock.NewResult(1, 1))
+	mock.ExpectExec("DELETE FROM memories.*deleted_at IS NULL").
+		WithArgs(key.MemoryID, key.AppName, key.UserID).
+		WillReturnResult(sqlmock.NewResult(0, 0))
+	mock.ExpectRollback()
+
+	result := &memory.UpdateResult{MemoryID: "unchanged"}
+	err := svc.UpdateMemory(
+		context.Background(),
+		key,
+		"target",
+		nil,
+		memory.WithUpdateResult(result),
+	)
+	require.ErrorContains(t, err, "not found")
+	require.Equal(t, "unchanged", result.MemoryID)
 }
